Require necessary files where they're needed
The previous commit's issue was hidden by the fact that in tests,
better_html/parser was required by test_helper, which made it
unnecessary to require in the right place.

This commit removes the require from the test_helper, and then adds all
the necessary requires.

To find all the missing requires, I used this to load and run one test
file at a time:

    $ find test -iname '*_test.rb' | xargs -tn1 bundle exec ruby -Ilib:test