Only pass template language to parser
`assert_erb_safety` accepts a `filename` option which prints the filename in the error message, which is invaluable for finding which file the error comes from (without this the error message just reports that there is an error in "(buffer)"). However, passing the filename right now throws an error: `ArgumentError: unknown keyword: :filename`, which comes about because `assert_erb_safety` passes all of it's own options to straight to `BetterHtml::Parser` and that doesn't support a `filename` option (in fact it only supports `template_language`).

I've changed `assert_erb_safety` so that it only passes the `template_language` option through to `Parser`, which fixes the error and allows `assert_erb_safety` to correctly include the filename in the assertion message.