Print the filename if set
If you run `ErubiImplementation#validate!` (as described in the readme) and there's a validation error it will print the line number and column but not the filename. You can pass the filename to erubi (like this `BetterHtml::BetterErb::ErubiImplementation.new(data, config: BetterHtml.config, filename:).validate!`) but `ErubiImplementation` doesn't use that information in it's error message.

I've updated runtime checks so that if the filename is passed to `ErubiImplementation` it will be used in error messages, which can really help when fixing errors.