Use Base64 to pass text through compiled erb templates
This is a proposed way to potentially reduce confusion in the error
highlighter in rails. The problem is partly that the highlighter looks
for the source erb expressions in the compiled template. The other big
problem is that multi-line blocks of code get double the newlines
added to the compiled source, causing the highlighter to fail to find
the correct line. This attempt to reduce the likelihood of those
issues.